requirements.txt
requests
numpy
faiss-cpu
streamlit